+++ src/App.tsx
@@ -8,24 +8,24 @@ import Search from './pages/Search';
 import NewArticle from './pages/NewArticle';
 import Catalog from './pages/Catalog';
 
-import ArticlesAvailables from './components/ArticlesAvailables';
+import ArticlesAvailables from './components/ui/ArticlesAvailableBySeller/ArticlesAvailables';
 
 function App() {
   return (
-      <BrowserRouter>
-        <Routes>
-          <Route path="/" element={<ArticlesAvailables />} />
-          {/* Page Achat */}
-          <Route path="/items/:id" element={<Article />} />
-          {/* Page vente si connecté */}
-          <Route path="/new" element={<NewArticle />} />
-          {/* Page de recherche global */}
-          <Route path="/search" element={<Search />} />
-          {/* Page de recherche par categorie via navbar */}
-          <Route path="/catalog/:id" element={<Catalog />} />
-          <Route path="/*" element={<Error404 />} />
-        </Routes>
-      </BrowserRouter>
+    <BrowserRouter>
+      <Routes>
+        <Route path="/" element={<ArticlesAvailables />} />
+        {/* Page Achat */}
+        <Route path="/items/:id" element={<Article />} />
+        {/* Page vente si connecté */}
+        <Route path="/new" element={<NewArticle />} />
+        {/* Page de recherche global */}
+        <Route path="/search" element={<Search />} />
+        {/* Page de recherche par categorie via navbar */}
+        <Route path="/catalog/:id" element={<Catalog />} />
+        <Route path="/*" element={<Error404 />} />
+      </Routes>
+    </BrowserRouter>
   );
 }
